+++ indico/modules/events/registration/client/js/form/fields/registry.js
@@ -10,15 +10,25 @@ import {getPluginObjects} from 'indico/utils/plugins';
 
 import AccommodationInput, {
   AccommodationSettings,
-  accommodationSettingsInitialData,
   accommodationSettingsFormValidator,
+  accommodationSettingsInitialData,
+  accommodationShowIfOptions,
+  accommodationGetDataForCondition,
 } from './AccommodationInput';
 import AccompanyingPersonsInput, {
   AccompanyingPersonsSettings,
   accompanyingPersonsSettingsInitialData,
 } from './AccompanyingPersonsInput';
-import BooleanInput, {BooleanSettings} from './BooleanInput';
-import CheckboxInput, {CheckboxSettings} from './CheckboxInput';
+import BooleanInput, {
+  BooleanSettings,
+  booleanShowIfOptions,
+  booleanGetDataForCondition,
+} from './BooleanInput';
+import CheckboxInput, {
+  CheckboxSettings,
+  checkboxShowIfOptions,
+  checkboxGetDataForCondition,
+} from './CheckboxInput';
 import {choiceFieldsSettingsFormDecorator} from './ChoicesSetup';
 import CountryInput, {CountrySettings} from './CountryInput';
 import DateInput, {
@@ -32,6 +42,8 @@ import LabelInput from './LabelInput';
 import MultiChoiceInput, {
   MultiChoiceSettings,
   multiChoiceSettingsInitialData,
+  multiChoiceShowIfOptions,
+  multiChoiceGetDataForCondition,
 } from './MultiChoiceInput';
 import NumberInput, {NumberSettings, numberSettingsFormValidator} from './NumberInput';
 import PhoneInput from './PhoneInput';
@@ -40,13 +52,15 @@ import SingleChoiceInput, {
   SingleChoiceSettings,
   singleChoiceSettingsFormDecorator,
   singleChoiceSettingsInitialData,
+  singleChoiceShowIfOptions,
+  singleChoiceGetDataForCondition,
 } from './SingleChoiceInput';
 import TextAreaInput, {TextAreaSettings} from './TextAreaInput';
 import TextInput, {TextSettings, textSettingsFormValidator} from './TextInput';
 import TimetableSessionsInput, {
+  sessionsSettingsFormValidator,
   TimetableSessionsSettings,
   timetableSessionsSettingsInitialData,
-  sessionsSettingsFormValidator,
 } from './TimetableSessions';
 
 /*
@@ -71,6 +85,11 @@ Available keys:
 - renderAsFieldset: optional; whether the field should be rendered in a fieldset; applies
   to fields that use multiple controls, like radios, checkboxes, multi-button controls;
   can either be a Boolean or a function that takes field options and returns a Boolean
+- showIfOptions: optional; a function that returns the dropdown options when selecting the
+  field as a condition
+- getDataForCondition: optional; a function that takes the current value of the field and
+  returns a list that can be compared with the stored condition of another field. its logic
+  must match the corresponding ``get_data_for_condition`` method in the backend field class
 */
 
 const fieldRegistry = {
@@ -112,6 +131,8 @@ const fieldRegistry = {
     hasPrice: true,
     icon: 'checkbox-checked',
     customFormItem: true,
+    showIfOptions: checkboxShowIfOptions,
+    getDataForCondition: checkboxGetDataForCondition,
   },
   date: {
     title: Translate.string('Date'),
@@ -128,6 +149,8 @@ const fieldRegistry = {
     hasPrice: true,
     icon: 'switchon',
     renderAsFieldset: true,
+    showIfOptions: booleanShowIfOptions,
+    getDataForCondition: booleanGetDataForCondition,
   },
   phone: {
     title: Translate.string('Phone'),
@@ -158,6 +181,8 @@ const fieldRegistry = {
     settingsModalSize: 'small',
     settingsFormDecorators: [choiceFieldsSettingsFormDecorator, singleChoiceSettingsFormDecorator],
     settingsFormInitialData: singleChoiceSettingsInitialData,
+    showIfOptions: singleChoiceShowIfOptions,
+    getDataForCondition: singleChoiceGetDataForCondition,
     icon: 'dropmenu',
     renderAsFieldset: options => options.itemType === 'radiogroup',
   },
@@ -168,6 +193,8 @@ const fieldRegistry = {
     settingsModalSize: 'small',
     settingsFormDecorators: [choiceFieldsSettingsFormDecorator],
     settingsFormInitialData: multiChoiceSettingsInitialData,
+    showIfOptions: multiChoiceShowIfOptions,
+    getDataForCondition: multiChoiceGetDataForCondition,
     icon: 'list',
     renderAsFieldset: true,
   },
@@ -183,6 +210,8 @@ const fieldRegistry = {
     alwaysRequired: true,
     icon: 'home',
     renderAsFieldset: true,
+    showIfOptions: accommodationShowIfOptions,
+    getDataForCondition: accommodationGetDataForCondition,
   },
   accompanying_persons: {
     title: Translate.string('Accompanying Persons'),