+++ indico/modules/events/registration/client/js/form/selectors.js
@@ -14,12 +14,28 @@ export const getStaticData = state => state.staticData;
 
 export const getFlatSections = state => state.sections;
 export const getItems = state => state.items;
+export const getHiddenItemsInitialized = state => state.hiddenItems.ready;
+export const getHiddenItemIds = state => state.hiddenItems.ids;
 
 /** Get an item by its ID. */
 export const getItemById = createSelector(
   getItems,
   (__, itemId) => itemId,
   (fields, itemId) => fields[itemId]
+);
+
+/** Check whether an item is hidden. */
+export const isItemHidden = createSelector(
+  getHiddenItemIds,
+  (__, itemId) => itemId,
+  (hiddenIds, itemId) => !!itemId && hiddenIds.includes(itemId)
+);
+
+/** Get the HTML names of hidden items. */
+export const getHiddenItemHTMLNames = createSelector(
+  getItems,
+  getHiddenItemIds,
+  (fields, hiddenIds) => Object.values(_.pick(fields, hiddenIds)).map(f => f.htmlName)
 );
 
 /** Get the currency used by the registration form. */