+++ indico/modules/events/registration/client/js/form_submission/RegistrationFormSubmission.jsx
@@ -26,9 +26,15 @@ import {indicoAxios} from 'indico/utils/axios';
 import {getPluginObjects, renderPluginComponents} from 'indico/utils/plugins';
 
 import ConsentToPublishDropdown from '../components/ConsentToPublishDropdown';
+import ConditionalFieldsController from '../form/ConditionalFieldsController';
 import FormErrorList from '../form/FormErrorList';
 import FormSection from '../form/FormSection';
-import {getNestedSections, getStaticData} from '../form/selectors';
+import {
+  getHiddenItemHTMLNames,
+  getHiddenItemsInitialized,
+  getNestedSections,
+  getStaticData,
+} from '../form/selectors';
 
 import {
   getUpdateMode,
@@ -145,11 +151,15 @@ export default function RegistrationFormSubmission() {
   const publishToParticipants = useSelector(getPublishToParticipants);
   const publishToPublic = useSelector(getPublishToPublic);
   const showConsentToPublish = !isManagement && publishToParticipants !== 'hide_all';
+  const hiddenItemHTMLNames = useSelector(getHiddenItemHTMLNames);
+  const hiddenItemsInitialized = useSelector(getHiddenItemsInitialized);
 
   const onSubmit = async (data, form) => {
     // There is no need to store whether the user accepted the privacy policy
     // as it is not possible to submit the form without accepting it.
     data = _.omit(data, 'agreed_to_privacy_policy');
+    // Do not send data for hidden fields; they are cleared server-side.
+    data = _.omit(data, hiddenItemHTMLNames);
     let resp;
     const formData = isUpdateMode ? getChangedValues(data, form, ['notify_user']) : data;
     try {
@@ -175,11 +185,18 @@ export default function RegistrationFormSubmission() {
     >
       {fprops => (
         <form onSubmit={fprops.handleSubmit}>
-          {renderPluginComponents('regformBeforeSections')}
-          {sections.map(section => (
-            <FormSection key={section.id} {...section} />
-          ))}
-          {renderPluginComponents('regformAfterSections')}
+          <ConditionalFieldsController />
+          {hiddenItemsInitialized && (
+            // avoid rendering the form until the conditional fields controller had the chance
+            // to run, to avoid initially-hidden fields from showing up for a short moment on load
+            <>
+              {renderPluginComponents('regformBeforeSections')}
+              {sections.map(section => (
+                <FormSection key={section.id} {...section} />
+              ))}
+              {renderPluginComponents('regformAfterSections')}
+            </>
+          )}
           <div>
             {isManagement && <ManagementOptions isUpdateMode={isUpdateMode} />}
             {showConsentToPublish && (