+++ indico/modules/events/registration/controllers/management/sections.py
@@ -6,15 +6,18 @@
 # LICENSE file for more details.
 
 from flask import jsonify, request, session
+from marshmallow import ValidationError
 from werkzeug.exceptions import BadRequest
 
 from indico.core.db import db
+from indico.core.errors import NoReportError
 from indico.modules.events.registration import logger
 from indico.modules.events.registration.controllers.management import RHManageRegFormBase
 from indico.modules.events.registration.models.items import RegistrationFormItemType, RegistrationFormSection
 from indico.modules.events.registration.util import get_flat_section_positions_setup_data, update_regform_item_positions
 from indico.modules.logs.models.entries import EventLogRealm, LogKind
 from indico.modules.logs.util import make_diff_log
+from indico.util.i18n import _
 from indico.web.util import jsonify_data
 
 
@@ -73,6 +76,16 @@ class RHRegistrationFormModifySection(RHManageRegFormSectionBase):
         if self.section.type == RegistrationFormItemType.section_pd and changes.get('is_manager_only'):
             raise BadRequest
         changes = self.section.populate_from_dict(changes)
+        if changes.get('is_manager_only') == (False, True):
+            # Check no conditional fields depend on this section if it is becoming manager-only now
+            critical_fields_ids = {field.id for field in self.section.children if not field.is_deleted}
+            for section in self.regform.sections:
+                if section.is_deleted:
+                    continue
+                fields_ids = {field.show_if_id for field in section.children
+                              if field.show_if_id is not None and not field.is_deleted}
+                if critical_fields_ids & fields_ids:
+                    raise ValidationError('Cannot make section manager-only due to conditional field relations')
         db.session.flush()
         changes = make_diff_log(changes, {
             'title': {'title': 'Title', 'type': 'string'},
@@ -93,8 +106,13 @@ class RHRegistrationFormToggleSection(RHManageRegFormSectionBase):
 
     def _process_POST(self):
         enabled = request.args.get('enable') == 'true'
-        if not enabled and self.section.type == RegistrationFormItemType.section_pd:
-            raise BadRequest
+        if not enabled:
+            if self.section.type == RegistrationFormItemType.section_pd:
+                raise BadRequest
+            if any(f.condition_for for f in self.section.children):
+                raise NoReportError.wrap_exc(
+                    BadRequest(_('Sections with fields used as conditional cannot be disabled'))
+                )
         self.section.is_enabled = enabled
         update_regform_item_positions(self.regform)
         db.session.flush()