+++ indico/modules/events/registration/fields/base.py
@@ -16,12 +16,29 @@ from indico.util.i18n import _
 from indico.util.marshmallow import not_empty
 
 
-class BillableFieldDataSchema(mm.Schema):
+class FieldSetupSchemaBase(mm.Schema):
+    """Base schema for configuring registration form fields.
+
+    Any global options that go into every field's JSON config data can be
+    added here; if the options are stored in a dedicated column, they should
+    go to `GeneralFieldDataSchema` instead.
+    """
+
+
+class BillableFieldDataSchema(FieldSetupSchemaBase):
+    """Base config schema for fields that can have a price."""
+
     price = fields.Float(load_default=0)
 
 
 class LimitedPlacesBillableFieldDataSchema(BillableFieldDataSchema):
+    """Base config schema for fields that can have a price and limited places."""
+
     places_limit = fields.Integer(load_default=0, validate=validate.Range(0))
+
+
+class LimitedPlacesBillableItemSchema(LimitedPlacesBillableFieldDataSchema):
+    """Base config schema for field items that can have a price and limited places."""
 
 
 class RegistrationFormFieldBase:
@@ -38,7 +55,7 @@ class RegistrationFormFieldBase:
     #: additional options for the marshmallow field
     mm_field_kwargs = {}
     #: the marshmallow base schema for configuring the field
-    setup_schema_base_cls = mm.Schema
+    setup_schema_base_cls = FieldSetupSchemaBase
     #: a dict with extra marshmallow fields to include in the setup schema
     setup_schema_fields = {}
     #: whether this field is associated with a file instead of normal data
@@ -47,6 +64,10 @@ class RegistrationFormFieldBase:
     is_invalid_field = False
     #: whether this field must not be "empty" (falsy value) if required
     not_empty_if_required = True
+    #: whether this field may be used as a condition; this is purely for validation,
+    #: on the frontend it depends solely on the config in the field registry whether
+    #: the field can be selected as a condition.
+    allow_condition = False
 
     def __init__(self, form_item):
         self.form_item = form_item
@@ -86,6 +107,19 @@ class RegistrationFormFieldBase:
         assert default is not NotImplemented
         return default
 
+    def get_data_for_condition(self, value) -> set:
+        """Convert field value to the format expected in a condition check.
+
+        If this field can be used for a condition, this method needs to return a
+        set of values (based on the `value` the field currently contains). If this
+        set intersects with the other field's configured condition, that field will
+        be visible.
+
+        The logic in this method must match the corresponding ``getDataForCondition``
+        function in the frontend field registry.
+        """
+        raise NotImplementedError('Field cannot be used as a condition')
+
     def get_validators(self, existing_registration):
         """Return a list of marshmallow validators for this field.
 
@@ -114,10 +148,10 @@ class RegistrationFormFieldBase:
         """
         return RegistrationData.data.op('#>>')('{}').in_(data_list)
 
-    def create_setup_schema(self):
+    def create_setup_schema(self, context=None):
         name = f'{type(self).__name__}SetupDataSchema'
         schema = self.setup_schema_base_cls.from_dict(self.setup_schema_fields, name=name)
-        return schema()
+        return schema(context=context)
 
     def create_mm_field(self, registration=None, override_required=False):
         """
@@ -245,7 +279,7 @@ class RegistrationFormBillableField(RegistrationFormFieldBase):
 
 
 class RegistrationFormBillableItemsField(RegistrationFormBillableField):
-    setup_schema_base_cls = mm.Schema
+    setup_schema_base_cls = FieldSetupSchemaBase
 
     @classmethod
     def process_field_data(cls, data, old_data=None, old_versioned_data=None):