+++ CHANGES.rst
@@ -24,6 +24,8 @@ Improvements
   (:pr:`6860`)
 - Use :data:`STATIC_SITE_STORAGE` for the temporary file from a material package
   (:pr:`6898`)
+- Implement conditional fields in registration forms (:issue:`1227`, :pr:`6678`,
+  thanks :user:`Moliholy, omegak, unconventionaldotdev`)
 
 Bugfixes
 ^^^^^^^^